+++ tasmota/tasmota_template.h
@@ -146,6 +146,8 @@ enum UserSelectablePins {
   GPIO_IEM3000_TX, GPIO_IEM3000_RX,    // IEM3000 Serial interface
   GPIO_ZIGBEE_RST,                     // Zigbee reset
   GPIO_DYP_RX,
+  GPIO_MIEL_HVAC_TX,                   // Mitsubishi Electric HVAC TX pin
+  GPIO_MIEL_HVAC_RX,                   // Mitsubishi Electric HVAC RX pin
   GPIO_SENSOR_END };
 
 enum ProgramSelectablePins {
@@ -248,7 +250,8 @@ const char kSensorNames[] PROGMEM =
   D_SENSOR_LMT01_PULSE "|"
   D_SENSOR_IEM3000_TX "|" D_SENSOR_IEM3000_RX "|"
   D_SENSOR_ZIGBEE_RST "|"
-  D_SENSOR_DYP_RX
+  D_SENSOR_DYP_RX "|"
+  D_SENSOR_MIEL_HVAC_TX "|" D_SENSOR_MIEL_HVAC_RX
   ;
 
 const char kSensorNamesFixed[] PROGMEM =
@@ -574,6 +577,10 @@ const uint16_t kGpioNiceList[] PROGMEM = {
 #ifdef USE_TELEINFO
   AGPIO(GPIO_TELEINFO_RX),
   AGPIO(GPIO_TELEINFO_ENABLE),
+#endif
+#ifdef USE_MIEL_HVAC
+  AGPIO(GPIO_MIEL_HVAC_TX),    // Mitsubishi Electric HVAC TX pin
+  AGPIO(GPIO_MIEL_HVAC_RX),    // Mitsubishi Electric HVAC RX pin
 #endif
 
 #ifdef ESP32