+++ tasmota/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove auto config update for all Friendlynames and Switchtopic from versions before 8.x
 - Change redesigning ESP8266 GPIO internal representation in line with ESP32
 - Change new IR Raw compact format (#9444)
+- Add optional support for Mitsubishi Electric HVAC by David Gwynne (#9237)
 
 ### 8.5.1 20201002